some basic cleanup, not compiling yet

+++ src/main/java/com/yksi7417/simulator/Trade.java
@@ -1,14 +1,14 @@
 package com.yksi7417.simulator;
 
 public class Trade {
-	int size; 
+	long size; 
 	double price;
-	public Trade(int size, double price) {
+	public Trade(long size, double price) {
 		super();
 		this.size = size;
 		this.price = price;
 	}
-	public int getSize() {
+	public long getSize() {
 		return size;
 	}
 	public double getPrice() {
